runner/python-portainer/requirements.txt
PyYAML==5.4.1
regex==2022.3.15
requests==2.25.1
